requirements.txt
asgiref==3.3.4
Django==3.2
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
